+++ bin/travis-ci/docker_build.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/with-contenv sh
+#!/bin/bash
 #
 # Copyright (c) 2018 SD Elements Inc.
 #
@@ -17,15 +17,6 @@
 
 set -eo pipefail
 
-echo "Starting PHP-FPM"
-
-# Make unix socket for nginx/php
-mkdir -p /var/run/php-fpm
-touch /var/run/php-fpm/www.sock
-chown -R www-data:www-data /var/run/php-fpm
-
-# Configure Moodle (it will wait for PHP/PostgreSQL in the background)
-(/usr/local/bin/configure-moodle.sh) &
-
-# Start PHP-FPM
-"/usr/sbin/php-fpm${PHP_VERSION}" -R -F
+# Build image
+docker-compose --file docker-compose.yml --file dc.build.yml build --no-cache nginx-php-moodle
+docker images