+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2018-09-04
+## [v0.1.3] - 2018-09-07
+### Added
+- Travis CI jobs
+  - Dockerfile linter
+  - Docker image build and publish to Cargo
+  - DEB package build
+  - SDLC script to ensure version bumps
+
+### Changed
+- Minor bug fixes for `dc.deb.yml`
+- Minor updates to the Moodle and DEB `Dockerfile`
+- Strict mode and copyright notice in bash scripts
+
+## [v0.1.2] - 2018-09-04
 ### Added
 - DEB Package
   - Added `iteration` to DEB name
@@ -18,7 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - Explicit DEB filename, `fpm` will create it using `version` and `iteration`
 
-## [0.1.1] - 2018-08-23
+## [v0.1.1] - 2018-08-23
 ### Added
 - `systemd` files to DEB package
   - `docker-moodle` unit file
@@ -31,7 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fetch images from Docker registry URL by default
 - Update `README.md` with latest changes
 
-## [0.1.0] - 2018-08-13
+## [v0.1.0] - 2018-08-13
 ### Added
 - Project `CHANGELOG.md`
 - Container for building DEB packages